requirements.txt
# Polymarket HFT Bot Dependencies
py-clob-client>=0.17.0
python-dotenv>=1.0.0
httpx>=0.25.0
websockets>=12.0
pydantic>=2.5.0
numpy>=1.26.0
scipy>=1.11.0
google-genai>=1.0.0
eth-account>=0.10.0
web3>=6.15.0
maturin>=1.4.0
